# docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: X-Pilot Agentic Server API
  version: 1.0.0
  description: |
    可直接导入 Apifox 的 OpenAPI 文档。
    注意：
    - 认证方式：`Authorization: Bearer <apiKey>`
    - 业务文稿接口路径是 `/api/business/document-generation/*`，不是 `document-generator`
servers:
  - url: http://192.168.50.54:7800
    description: Local Network Dev
security:
  - BearerAuth: []
...
paths:
...
  /api/agents:
    get:
      tags: [Agents]
      summary: 获取 Agent 列表
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                - name: document-generator
                  model: z-ai/glm-5
                  tools: ["*"]
                  skills: ["self-reflect", "self-evolve", "memory-search"]
                  heartbeatEnabled: true
                  evolutionEnabled: true
    post:
      tags: [Agents]
      summary: 动态创建 Agent
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: custom-agent
              identity: "你是一个课程大纲助手。"
              mission: "输出结构化教学提纲。"
              soul: "保持严谨、可执行。"
              group: system
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              example:
                name: custom-agent
                config:
                  model: z-ai/glm-5
                skills: ["self-reflect"]
        "400":
          description: 参数错误
          content:
            application/json:
              example:
                error: "Agent already exists: custom-agent"
  /api/agents/{name}:
    get:
      tags: [Agents]
      summary: 获取单个 Agent 详情
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                name: document-generator
                config:
                  model: z-ai/glm-5
                  maxConcurrency: 5
                prompts:
                  identity: "..."
                  mission: "..."
                skills:
                  - name: self-reflect
                    description: built-in
                memoryPath: "apps/document-generation/document-generator/memory"
        "404":
          description: 未找到
          content:
            application/json:
              example: { error: "Agent not found" }
  /api/agents/{name}/invoke:
    post:
      tags: [Agents]
      summary: 同步调用 Agent
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      requestBody:
        required: true
        content:
          application/json:
            example:
              input: "请生成《电工安全实训》教学文稿"
              threadId: "thread-1"
              sessionId: "test0321"
              context:
                businessType: document-generation
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                taskId: "task_abc123"
                threadId: "thread-1"
                output: "已完成文稿生成。"
        "404":
          description: Agent 不存在
          content:
            application/json:
              example: { error: "Agent not found" }
  /api/agents/{name}/invoke-async:
    post:
      tags: [Agents]
      summary: 异步调用 Agent
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      requestBody:
        required: true
        content:
          application/json:
            example:
              input: "生成文稿"
              sessionId: "test0321"
              webhookUrl: "https://example.com/webhook"
      responses:
        "202":
          description: 已受理
          content:
            application/json:
              example:
                taskId: "task_async_001"
                status: "pending"
        "500":
          description: 服务错误
          content:
            application/json:
              example: { error: "Async agent invocation failed" }
  /api/agents/{name}/stream:
    post:
      tags: [Agents]
      summary: SSE 流式调用 Agent（v2）
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      requestBody:
        required: true
        content:
          application/json:
            example:
              input: "流式生成文稿"
              sessionId: "test0321"
      responses:
        "200":
          description: SSE 事件流
          content:
            text/event-stream:
              example: |
                event: message
                data: {"event":"message","data":"开始生成..."}
                
                event: task_finished
                data: {"event":"task_finished","task_id":"task_1"}
  /api/agents/{name}/reload:
    post:
      tags: [Agents]
      summary: 重载 Agent
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                name: document-generator
                config:
                  model: z-ai/glm-5
  /api/agents/{name}/memory:
    get:
      tags: [Agents]
      summary: 查询 Agent 记忆
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
        - in: query
          name: query
          required: false
          schema: { type: string }
          example: 电工实训
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                results:
                  - content: "历史经验：先提纲后成文"
                    score: 0.91
    post:
      tags: [Agents]
      summary: 追加 Agent 记忆
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      requestBody:
        required: true
        content:
          application/json:
            example:
              content: "本次模板只用于结构，不可复用原文。"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example: { success: true }
  /api/agents/{name}/evolution:
    get:
      tags: [Agents]
      summary: 获取待审批进化提案
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                proposals:
                  - id: evo_001
                    reason: "提升提纲鲁棒性"
  /api/agents/{name}/evolution/{id}/approve:
    post:
      tags: [Agents]
      summary: 审批通过进化提案
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
          example: evo_001
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                success: true
                proposal:
                  id: evo_001
  /api/agents/{name}/evolution/{id}/reject:
    post:
      tags: [Agents]
      summary: 拒绝进化提案
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: path
          name: id
          required: true
          schema: { type: string }
          example: evo_001
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                success: true
                proposal:
                  id: evo_001
...
  /api/sessions/{sessionId}/events:
    get:
      tags: [Sessions]
      summary: 会话事件流（SSE）
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
          example: test0321
      responses:
        "200":
          description: SSE
          content:
            text/event-stream:
              example: |
                data: {"event":"task_started","sessionId":"test0321"}
  /api/sessions/{sessionId}/artifacts:
    get:
      tags: [Sessions]
      summary: 获取会话产物列表
      parameters:
        - in: path
          name: sessionId
          required: true
          schema: { type: string }
          example: test0321
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                sessionId: test0321
                artifacts:
                  - name: artifacts/document.md
                    size: 12456
                    createdAt: "2026-03-19T03:30:00.000Z"
                    modifiedAt: "2026-03-19T03:32:00.000Z"
...
  /ws/agents/{name}:
    get:
      tags: [Agents]
      summary: WebSocket 流式调用 Agent
      description: |
        WebSocket 端点。连接后发送 JSON：
        `{"input":"...", "threadId":"...", "sessionId":"...", "context":{}}`
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
          example: document-generator
      responses:
        "101":
          description: WebSocket 握手成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key